replace datetime with datetimehelper

+++ src/BjBygg.Application/Application/Queries/DbSyncQueries/Common/DbSyncQuery.cs
@@ -5,6 +5,6 @@ namespace BjBygg.Application.Application.Queries.DbSyncQueries.Common
     public abstract class DbSyncQuery<T> : IRequest<DbSyncResponse<T>> where T : DbSyncDto
     {
         public int? InitialNumberOfMonths { get; set; }
-        public double? Timestamp { get; set; }
+        public long? Timestamp { get; set; }
     }
 }
