+++ src/BjBygg.Application/Application/Queries/DbSyncQueries/SyncAll/SyncAllQuery.cs
@@ -6,21 +6,21 @@ namespace BjBygg.Application.Application.Queries.DbSyncQueries.SyncAll
     {
         public int? InitialNumberOfMonths { get; set; }
 
-        public double? MissionTimestamp { get; set; }
+        public long? MissionTimestamp { get; set; }
 
-        public double? EmployerTimestamp { get; set; }
+        public long? EmployerTimestamp { get; set; }
 
-        public double? MissionImageTimestamp { get; set; }
+        public long? MissionImageTimestamp { get; set; }
 
-        public double? MissionNoteTimestamp { get; set; }
+        public long? MissionNoteTimestamp { get; set; }
 
-        public double? MissionDocumentTimestamp { get; set; }
+        public long? MissionDocumentTimestamp { get; set; }
 
-        public double? DocumentTypeTimestamp { get; set; }
+        public long? DocumentTypeTimestamp { get; set; }
 
-        public double? MissionTypeTimestamp { get; set; }
+        public long? MissionTypeTimestamp { get; set; }
 
-        public double? UserTimesheetTimestamp { get; set; }
+        public long? UserTimesheetTimestamp { get; set; }
 
     }
 }
